Reorganizando Arquitetura Clean + Ajustes
