+++ src/main/java/br/com/alura/escola/VOUtils/CPF.java
@@ -1,0 +1,20 @@
+package br.com.alura.escola.VOUtils;
+
+//valueObject
+public class CPF {
+
+	private String numero;
+
+	public CPF(String numero) {
+		if (numero == null || 
+				!numero.matches("\\d{3}\\.\\d{3}\\.\\d{3}\\-\\d{2}")) {
+			throw new IllegalArgumentException("CPF invalido!");
+		}
+		this.numero = numero;
+	}
+	
+	public String getNumero() {
+		return numero;
+	}
+	
+}
