+++ src/main/java/br/com/alura/escola/VOUtils/Email.java
@@ -1,0 +1,21 @@
+package br.com.alura.escola.VOUtils;
+
+//valueObject
+public class Email {
+	
+	private String endereco;
+
+	public Email(String endereco) {
+		if (endereco == null || 
+				!endereco.matches("^[a-zA-Z0-9._]+@[a-zA-Z0-9.-]+\\.[a-zA-Z]{2,}$")) {
+			throw new IllegalArgumentException("E-mail invalido!");
+		}
+
+		this.endereco = endereco;
+	}
+	
+	public String getEndereco() {
+		return endereco;
+	}
+
+}
