--- src/main/java/br/com/alura/escola/aplicacao/aluno/matricular/MatricularAluno.java
@@ -1,20 +1,0 @@
-package br.com.alura.escola.aplicacao.aluno.matricular;
-
-import br.com.alura.escola.dominio.aluno.Aluno;
-import br.com.alura.escola.dominio.aluno.RepositorioDeAlunos;
-
-public class MatricularAluno {
-	
-	private final RepositorioDeAlunos repositorio;
-
-	public MatricularAluno(RepositorioDeAlunos repositorio) {
-		this.repositorio = repositorio;
-	}
-	
-	// COMMAND
-	public void executa(MatricularAlunoDto dados) {
-		Aluno novo = dados.criarAluno();
-		repositorio.matricular(novo);
-	}
-
-}
