--- src/main/java/br/com/alura/escola/aplicacao/indicacao/EnviarEmailIndicacao.java
@@ -1,9 +1,0 @@
-package br.com.alura.escola.aplicacao.indicacao;
-
-import br.com.alura.escola.dominio.aluno.Aluno;
-
-public interface EnviarEmailIndicacao {
-	
-	void enviarPara(Aluno indicado);
-
-}
