+++ src/main/java/br/com/alura/escola/domain/aluno/Aluno.java
@@ -1,0 +1,56 @@
+package br.com.alura.escola.domain.aluno;
+
+import br.com.alura.escola.VOUtils.CPF;
+import br.com.alura.escola.VOUtils.Email;
+import br.com.alura.escola.VOUtils.Telefone;
+import java.util.ArrayList;
+import java.util.List;
+
+public class Aluno {
+	
+	private CPF cpf;
+	private String nome;
+	
+	private Email email;
+	
+	private List<Telefone> telefones = new ArrayList<>();
+	
+	private String senha;
+	
+	public Aluno(CPF cpf, String nome, Email email) {
+		this.cpf = cpf;
+		this.nome = nome;
+		this.email = email;
+	}
+
+	public void adicionarTelefone(String ddd, String numero) {
+		this.telefones.add(new Telefone(ddd, numero));
+	}
+	
+	public String getCpf() {
+		return cpf.getNumero();
+	}
+	
+	public String getNome() {
+		return nome;
+	}
+	
+	public String getEmail() {
+		return email.getEndereco();
+	}
+	
+	public List<Telefone> getTelefones() {
+		return telefones;
+	}
+
+	public String getSenha() {
+		return senha;
+	}
+	
+}
+
+
+
+
+
+
