+++ src/main/java/br/com/alura/escola/domain/aluno/AlunoNaoEncontrado.java
@@ -1,0 +1,13 @@
+package br.com.alura.escola.domain.aluno;
+
+import br.com.alura.escola.VOUtils.CPF;
+
+public class AlunoNaoEncontrado extends RuntimeException {
+
+	private static final long serialVersionUID = 1L;
+
+	public AlunoNaoEncontrado(CPF cpf) {
+		super("Aluno nao encontrado com CPF: " + cpf.getNumero());
+	}
+
+}
