+++ src/main/java/br/com/alura/escola/domain/auth/CifradorDeSenha.java
@@ -1,0 +1,8 @@
+package br.com.alura.escola.domain.auth;
+
+public interface CifradorDeSenha {
+
+	String cifrarSenha(String senha);
+	
+	boolean validarSenhaCifrada(String senhaCifrada, String senha);
+}
