--- src/main/java/br/com/alura/escola/dominio/aluno/AlunoNaoEncontrado.java
@@ -1,11 +1,0 @@
-package br.com.alura.escola.dominio.aluno;
-
-public class AlunoNaoEncontrado extends RuntimeException {
-
-	private static final long serialVersionUID = 1L;
-
-	public AlunoNaoEncontrado(CPF cpf) {
-		super("Aluno nao encontrado com CPF: " + cpf.getNumero());
-	}
-
-}
