--- src/main/java/br/com/alura/escola/dominio/aluno/CifradorDeSenha.java
@@ -1,8 +1,0 @@
-package br.com.alura.escola.dominio.aluno;
-
-public interface CifradorDeSenha {
-
-	String cifrarSenha(String senha);
-	
-	boolean validarSenhaCifrada(String senhaCifrada, String senha);
-}
