--- src/main/java/br/com/alura/escola/dominio/aluno/RepositorioDeAlunos.java
@@ -1,15 +1,0 @@
-package br.com.alura.escola.dominio.aluno;
-
-import java.util.List;
-
-public interface RepositorioDeAlunos {
-	
-	void matricular(Aluno aluno);
-	
-	Aluno buscarPorCPF(CPF cpf);
-	
-	List<Aluno> listarTodosAlunosMatriculados();
-	
-	//...
-
-}
