--- src/main/java/br/com/alura/escola/dominio/aluno/Telefone.java
@@ -1,32 +1,0 @@
-package br.com.alura.escola.dominio.aluno;
-
-public class Telefone {
-	
-	private String ddd;
-	private String numero;
-
-	public Telefone(String ddd, String numero) {
-		if (ddd == null || numero == null) {
-			throw new IllegalArgumentException("DDD e Numero sao obrigatorios!");
-		}
-
-		if (!ddd.matches("\\d{2}")) {
-			throw new IllegalArgumentException("DDD invalido!");
-		}
-		
-		if (!numero.matches("\\d{8}|\\d{9}")) {
-			throw new IllegalArgumentException("Numero invalido!");
-		}
-		this.ddd = ddd;
-		this.numero = numero;
-	}
-
-	public String getDdd() {
-		return ddd;
-	}
-	
-	public String getNumero() {
-		return numero;
-	}
-
-}
