+++ src/main/java/br/com/alura/escola/external/aluno/RepositorioDeAlunosEmMemoria.java
@@ -1,0 +1,33 @@
+package br.com.alura.escola.external.aluno;
+
+import java.util.ArrayList;
+import java.util.List;
+
+import br.com.alura.escola.domain.aluno.Aluno;
+import br.com.alura.escola.domain.aluno.AlunoNaoEncontrado;
+import br.com.alura.escola.VOUtils.CPF;
+import br.com.alura.escola.domain.aluno.RepositorioDeAlunos;
+
+public class RepositorioDeAlunosEmMemoria implements RepositorioDeAlunos {
+
+	private List<Aluno> matriculados = new ArrayList<>();
+	
+	@Override
+	public void matricular(Aluno aluno) {
+		this.matriculados.add(aluno);
+	}
+
+	@Override
+	public Aluno buscarPorCPF(CPF cpf) {
+		return this.matriculados.stream()
+				.filter(a -> a.getCpf().equals(cpf.getNumero()))
+				.findFirst()
+				.orElseThrow(() -> new AlunoNaoEncontrado(cpf));
+	}
+
+	@Override
+	public List<Aluno> listarTodosAlunosMatriculados() {
+		return this.matriculados;
+	}
+
+}
