--- src/main/java/br/com/alura/escola/infra/indicacao/EnviarEmailIndicacaoComJavaMail.java
@@ -1,12 +1,0 @@
-package br.com.alura.escola.infra.indicacao;
-
-import br.com.alura.escola.aplicacao.indicacao.EnviarEmailIndicacao;
-import br.com.alura.escola.dominio.aluno.Aluno;
-
-public class EnviarEmailIndicacaoComJavaMail implements EnviarEmailIndicacao {
-
-	@Override
-	public void enviarPara(Aluno indicado) {
-		// logica de envio de email com a lib Java Mail
-	}
-}
