+++ src/main/java/br/com/alura/escola/useCases/aluno/matricular/MatricularAluno.java
@@ -1,0 +1,20 @@
+package br.com.alura.escola.useCases.aluno.matricular;
+
+import br.com.alura.escola.domain.aluno.Aluno;
+import br.com.alura.escola.domain.aluno.RepositorioDeAlunos;
+
+public class MatricularAluno {
+	
+	private final RepositorioDeAlunos repositorio;
+
+	public MatricularAluno(RepositorioDeAlunos repositorio) {
+		this.repositorio = repositorio;
+	}
+	
+	// COMMAND
+	public void executa(MatricularAlunoDto dados) {
+		Aluno novo = dados.criarAluno();
+		repositorio.matricular(novo);
+	}
+
+}
