+++ src/main/java/br/com/alura/escola/useCases/indicacao/EnviarEmailIndicacao.java
@@ -1,0 +1,9 @@
+package br.com.alura.escola.useCases.indicacao;
+
+import br.com.alura.escola.domain.aluno.Aluno;
+
+public interface EnviarEmailIndicacao {
+	
+	void enviarPara(Aluno indicado);
+
+}
