--- src/test/java/br/com/alura/escola/dominio/aluno/CPFTest.java
@@ -1,28 +1,0 @@
-package br.com.alura.escola.dominio.aluno;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-import org.junit.jupiter.api.Test;
-
-class CPFTest {
-
-	@Test
-	void naoDeveriaCriarCPFComNumerosInvalidos() {
-		assertThrows(IllegalArgumentException.class,
-				() -> new CPF(null));
-		
-		assertThrows(IllegalArgumentException.class,
-				() -> new CPF(""));
-		
-		assertThrows(IllegalArgumentException.class,
-				() -> new CPF("12345678900"));
-	}
-
-	@Test
-	void deveriaPermitirCriarCPFComNumeroValido() {
-		String numero = "123.456.789-00";
-		CPF cpf = new CPF(numero);
-		assertEquals(numero, cpf.getNumero());
-	}
-
-}
