--- src/test/java/br/com/alura/escola/dominio/aluno/EmailTest.java
@@ -1,30 +1,0 @@
-package br.com.alura.escola.dominio.aluno;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-import org.junit.jupiter.api.Test;
-
-import br.com.alura.escola.dominio.aluno.Email;
-
-class EmailTest {
-
-	@Test
-	void naoDeveriaCriarEmailsComEnderecosInvalidos() {
-		assertThrows(IllegalArgumentException.class,
-				() -> new Email(null));
-		
-		assertThrows(IllegalArgumentException.class,
-				() -> new Email(""));
-		
-		assertThrows(IllegalArgumentException.class,
-				() -> new Email("emailinvalido"));
-	}
-
-	@Test
-	void deveriaPermitirCriarEmailComEnderecoValido() {
-		String endereco = "fulano@alura.com.br";
-		Email email = new Email(endereco);
-		assertEquals(endereco, email.getEndereco());
-	}
-
-}
